chore: unify StaticInferencePool and ElasticInferencePool into single InferencePool
Consolidates the two pool implementations into one class that handles both
static (fixed base_url) and elastic (DNS discovery) modes. Adds separate
eval clients using openai_chat_completions (non-token) while training
defaults to openai_chat_completions_token.

- Merge StaticInferencePool + ElasticInferencePool → InferencePool
- Add eval_clients property and get_eval_client() for eval rollouts
- Slim elastic.py down to discovery functions and data classes
- Remove InferencePool protocol (single implementation)
- Update orchestrator to pass train/eval client types separately
