+++ tests/unit/utils/test_elastic.py
@@ -5,9 +5,10 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import httpx
 
+from prime_rl.configs.shared import ClientConfig, ElasticConfig
+from prime_rl.utils.client import InferencePool
 from prime_rl.utils.elastic import (
     AdapterState,
-    ElasticInferencePool,
     check_server_model,
     discover_ready_servers,
     discover_server_ips,
@@ -180,84 +181,67 @@ def test_adapter_state_creation():
     assert adapter.step == 100
 
 
-# ElasticInferencePool adapter matching tests
+# InferencePool adapter matching tests
+
+
+def _make_elastic_pool(**overrides) -> InferencePool:
+    """Helper to create an InferencePool in elastic mode for testing."""
+    config = ClientConfig(
+        elastic=ElasticConfig(hostname="test.hostname"),
+        base_url=["http://localhost:8000/v1"],
+    )
+    with patch("prime_rl.utils.client.get_logger"):
+        return InferencePool(
+            client_config=config,
+            model_name=overrides.get("model_name", "base-model"),
+        )
 
 
 def test_adapter_matches_when_no_adapter_desired():
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
-        # No adapter desired (base model inference)
-        assert pool._adapter_matches_desired(None) is True
-        assert pool._adapter_matches_desired(AdapterState("x", Path("/x"), 0)) is True
+    pool = _make_elastic_pool()
+    # No adapter desired (base model inference)
+    assert pool._adapter_matches_desired(None) is True
+    assert pool._adapter_matches_desired(AdapterState("x", Path("/x"), 0)) is True
 
 
 def test_adapter_matches_by_path():
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
-        pool._desired.path = Path("/weights/step_100")
-        pool._desired.step = 100
+    pool = _make_elastic_pool()
+    pool._desired.path = Path("/weights/step_100")
+    pool._desired.step = 100
 
-        loaded = AdapterState(name="lora", path=Path("/weights/step_100"), step=100)
-        assert pool._adapter_matches_desired(loaded) is True
+    loaded = AdapterState(name="lora", path=Path("/weights/step_100"), step=100)
+    assert pool._adapter_matches_desired(loaded) is True
 
-        loaded_wrong_path = AdapterState(name="lora", path=Path("/weights/step_50"), step=50)
-        assert pool._adapter_matches_desired(loaded_wrong_path) is False
+    loaded_wrong_path = AdapterState(name="lora", path=Path("/weights/step_50"), step=50)
+    assert pool._adapter_matches_desired(loaded_wrong_path) is False
 
 
 def test_adapter_matches_by_step_when_nonzero():
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
-        pool._desired.path = Path("/weights/step_100")
-        pool._desired.step = 100
+    pool = _make_elastic_pool()
+    pool._desired.path = Path("/weights/step_100")
+    pool._desired.step = 100
 
-        # Different path but same step
-        loaded = AdapterState(name="lora", path=Path("/other/path"), step=100)
-        assert pool._adapter_matches_desired(loaded) is True
+    # Different path but same step
+    loaded = AdapterState(name="lora", path=Path("/other/path"), step=100)
+    assert pool._adapter_matches_desired(loaded) is True
 
 
 def test_adapter_does_not_match_by_zero_step():
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
-        pool._desired.path = Path("/weights/step_0")
-        pool._desired.step = 0
+    pool = _make_elastic_pool()
+    pool._desired.path = Path("/weights/step_0")
+    pool._desired.step = 0
 
-        # Step 0 should not match by step alone (avoid false positives)
-        loaded = AdapterState(name="lora", path=Path("/other/path"), step=0)
-        assert pool._adapter_matches_desired(loaded) is False
+    # Step 0 should not match by step alone (avoid false positives)
+    loaded = AdapterState(name="lora", path=Path("/other/path"), step=0)
+    assert pool._adapter_matches_desired(loaded) is False
 
 
 def test_adapter_returns_false_when_no_adapter_loaded():
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
-        pool._desired.path = Path("/weights/step_100")
-        pool._desired.step = 100
+    pool = _make_elastic_pool()
+    pool._desired.path = Path("/weights/step_100")
+    pool._desired.step = 100
 
-        assert pool._adapter_matches_desired(None) is False
+    assert pool._adapter_matches_desired(None) is False
 
 
 # _get_loaded_adapter tests
@@ -265,139 +249,115 @@ def test_adapter_returns_false_when_no_adapter_loaded():
 
 def test_get_loaded_adapter_finds_correct_adapter_when_multiple_loaded():
     """Test that _get_loaded_adapter returns the adapter matching desired name, not the first one."""
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
+    pool = _make_elastic_pool()
 
-        # Set the desired adapter name
-        pool._desired.name = "rft-target-run"
-        pool._desired.path = Path("/data/outputs/target_run/broadcasts/step_10")
-        pool._desired.step = 10
+    # Set the desired adapter name
+    pool._desired.name = "rft-target-run"
+    pool._desired.path = Path("/data/outputs/target_run/broadcasts/step_10")
+    pool._desired.step = 10
 
-        # Mock admin client with multiple adapters (like the real scenario)
-        mock_admin = AsyncMock()
-        mock_response = MagicMock()
-        mock_response.raise_for_status = MagicMock()
-        # Simulates vLLM response with multiple adapters from different runs
-        mock_response.json.return_value = {
-            "data": [
-                {
-                    "id": "base-model",
-                    "parent": None,
-                    "root": "base-model",
-                },
-                {
-                    "id": "rft-other-run",  # Different run's adapter (comes first!)
-                    "parent": "base-model",
-                    "root": "/data/outputs/other_run/broadcasts/step_48",
-                },
-                {
-                    "id": "rft-target-run",  # Our desired adapter
-                    "parent": "base-model",
-                    "root": "/data/outputs/target_run/broadcasts/step_10",
-                },
-            ]
-        }
-        mock_admin.get.return_value = mock_response
-        pool._admin_clients["10.0.0.1"] = mock_admin
+    # Mock admin client with multiple adapters (like the real scenario)
+    mock_admin = AsyncMock()
+    mock_response = MagicMock()
+    mock_response.raise_for_status = MagicMock()
+    # Simulates vLLM response with multiple adapters from different runs
+    mock_response.json.return_value = {
+        "data": [
+            {
+                "id": "base-model",
+                "parent": None,
+                "root": "base-model",
+            },
+            {
+                "id": "rft-other-run",  # Different run's adapter (comes first!)
+                "parent": "base-model",
+                "root": "/data/outputs/other_run/broadcasts/step_48",
+            },
+            {
+                "id": "rft-target-run",  # Our desired adapter
+                "parent": "base-model",
+                "root": "/data/outputs/target_run/broadcasts/step_10",
+            },
+        ]
+    }
+    mock_admin.get.return_value = mock_response
+    pool._elastic_admin_clients["10.0.0.1"] = mock_admin
 
-        result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
+    result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
 
-        # Should return our target adapter, not the first one found
-        assert result is not None
-        assert result.name == "rft-target-run"
-        assert result.path == Path("/data/outputs/target_run/broadcasts/step_10")
-        assert result.step == 10
+    # Should return our target adapter, not the first one found
+    assert result is not None
+    assert result.name == "rft-target-run"
+    assert result.path == Path("/data/outputs/target_run/broadcasts/step_10")
+    assert result.step == 10
 
 
 def test_get_loaded_adapter_returns_none_when_desired_adapter_not_found():
     """Test that _get_loaded_adapter returns None when desired adapter is not in the list."""
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
+    pool = _make_elastic_pool()
 
-        pool._desired.name = "rft-missing-run"
-        pool._desired.path = Path("/data/outputs/missing_run/broadcasts/step_5")
-        pool._desired.step = 5
+    pool._desired.name = "rft-missing-run"
+    pool._desired.path = Path("/data/outputs/missing_run/broadcasts/step_5")
+    pool._desired.step = 5
 
-        mock_admin = AsyncMock()
-        mock_response = MagicMock()
-        mock_response.raise_for_status = MagicMock()
-        mock_response.json.return_value = {
-            "data": [
-                {"id": "base-model", "parent": None, "root": "base-model"},
-                {"id": "rft-other-run", "parent": "base-model", "root": "/data/outputs/other_run/step_10"},
-            ]
-        }
-        mock_admin.get.return_value = mock_response
-        pool._admin_clients["10.0.0.1"] = mock_admin
+    mock_admin = AsyncMock()
+    mock_response = MagicMock()
+    mock_response.raise_for_status = MagicMock()
+    mock_response.json.return_value = {
+        "data": [
+            {"id": "base-model", "parent": None, "root": "base-model"},
+            {"id": "rft-other-run", "parent": "base-model", "root": "/data/outputs/other_run/step_10"},
+        ]
+    }
+    mock_admin.get.return_value = mock_response
+    pool._elastic_admin_clients["10.0.0.1"] = mock_admin
 
-        result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
+    result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
 
-        assert result is None
+    assert result is None
 
 
 def test_get_loaded_adapter_parses_step_from_path():
     """Test that _get_loaded_adapter correctly parses step number from path."""
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
+    pool = _make_elastic_pool()
 
-        pool._desired.name = "my-lora"
+    pool._desired.name = "my-lora"
 
-        mock_admin = AsyncMock()
-        mock_response = MagicMock()
-        mock_response.raise_for_status = MagicMock()
-        mock_response.json.return_value = {
-            "data": [
-                {"id": "my-lora", "parent": "base", "root": "/weights/step_42"},
-            ]
-        }
-        mock_admin.get.return_value = mock_response
-        pool._admin_clients["10.0.0.1"] = mock_admin
+    mock_admin = AsyncMock()
+    mock_response = MagicMock()
+    mock_response.raise_for_status = MagicMock()
+    mock_response.json.return_value = {
+        "data": [
+            {"id": "my-lora", "parent": "base", "root": "/weights/step_42"},
+        ]
+    }
+    mock_admin.get.return_value = mock_response
+    pool._elastic_admin_clients["10.0.0.1"] = mock_admin
 
-        result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
+    result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
 
-        assert result is not None
-        assert result.step == 42
+    assert result is not None
+    assert result.step == 42
 
 
 def test_get_loaded_adapter_handles_step_dash_format():
     """Test that _get_loaded_adapter parses step-N format (with dash)."""
-    with patch("prime_rl.utils.elastic.get_logger"):
-        pool = ElasticInferencePool(
-            hostname="test.hostname",
-            client_config=MagicMock(),
-            model_name="base-model",
-            port=8000,
-        )
+    pool = _make_elastic_pool()
 
-        pool._desired.name = "my-lora"
+    pool._desired.name = "my-lora"
 
-        mock_admin = AsyncMock()
-        mock_response = MagicMock()
-        mock_response.raise_for_status = MagicMock()
-        mock_response.json.return_value = {
-            "data": [
-                {"id": "my-lora", "parent": "base", "root": "/weights/step-99"},
-            ]
-        }
-        mock_admin.get.return_value = mock_response
-        pool._admin_clients["10.0.0.1"] = mock_admin
+    mock_admin = AsyncMock()
+    mock_response = MagicMock()
+    mock_response.raise_for_status = MagicMock()
+    mock_response.json.return_value = {
+        "data": [
+            {"id": "my-lora", "parent": "base", "root": "/weights/step-99"},
+        ]
+    }
+    mock_admin.get.return_value = mock_response
+    pool._elastic_admin_clients["10.0.0.1"] = mock_admin
 
-        result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
+    result = asyncio.run(pool._get_loaded_adapter("10.0.0.1"))
 
-        assert result is not None
-        assert result.step == 99
+    assert result is not None
+    assert result.step == 99
